feat(utils): refactor utility imports and create utility functions
