+++ src/shared/components/ui/button.tsx
@@ -1,7 +1,7 @@
-import * as React from 'react'
 import { Slot } from '@radix-ui/react-slot'
 import { cva, type VariantProps } from 'class-variance-authority'
-import { cn } from '@lib/utils'
+import * as React from 'react'
+import { cn } from 'utils'
 
 const buttonVariants = cva(
   'inline-flex items-center justify-center whitespace-nowrap text-sm rounded-full lg:text-xl font-mono font-medium transition-colors focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:pointer-events-none disabled:opacity-50',
