+++ src/shared/components/ui/carousel.tsx
@@ -6,7 +6,7 @@ import useEmblaCarousel, {
   type UseEmblaCarouselType,
 } from 'embla-carousel-react'
 
-import { cn } from '@lib/utils'
+import { cn } from 'utils'
 import { Button } from './button'
 
 type CarouselApi = UseEmblaCarouselType[1]
