+++ src/shared/components/ui/sheet.tsx
@@ -3,7 +3,7 @@
 import * as React from 'react'
 import * as SheetPrimitive from '@radix-ui/react-dialog'
 import { cva, type VariantProps } from 'class-variance-authority'
-import { cn } from '@lib/utils'
+import { cn } from 'utils'
 import { Cross2Icon } from '@radix-ui/react-icons'
 
 const Sheet = SheetPrimitive.Root
